(chore) move cloud algorithm to a single file
